WidgerViewModel
