+++ src/WeatherTwentyOne/ViewModels/Page/FavoritesPageViewModel.cs
