+++ src/WeatherTwentyOne/ViewModels/Page/SettingsPageViewModel.cs
