+++ src/WeatherTwentyOne/ViewModels/Widget/WidgetViewModel.cs
@@ -1,0 +1,5 @@
+﻿namespace WeatherTwentyOne.ViewModels.Widget;
+
+public abstract class WidgetViewModel : BindableBase
+{
+}
