+++ src/WeatherTwentyOne/Views/Next24HrWidget.xaml.cs
@@ -1,24 +1,9 @@
-﻿using WeatherTwentyOne.ViewModels;
-
-namespace WeatherTwentyOne.Views;
+﻿namespace WeatherTwentyOne.Views;
 
 public partial class Next24HrWidget
 {
-    private HomePageViewModel viewModel;
-
     public Next24HrWidget()
     {
         InitializeComponent();
-
-        viewModel = ServiceLocator.GetService<HomePageViewModel>();
-
-        BindingContext = viewModel;
-    }
-
-    protected override async void OnBindingContextChanged()
-    {
-        base.OnBindingContextChanged();
-
-        await viewModel.InitializeAsync();
     }
 }
