+++ src/WeatherTwentyOne/Views/Next7DWidget.xaml.cs
@@ -1,24 +1,9 @@
-﻿using WeatherTwentyOne.ViewModels;
-
-namespace WeatherTwentyOne.Views;
+﻿namespace WeatherTwentyOne.Views;
 
 public partial class Next7DWidget
 {
-    private HomePageViewModel viewModel;
-
     public Next7DWidget()
     {
-        InitializeComponent();
-
-        viewModel = ServiceLocator.GetService<HomePageViewModel>();
-
-        BindingContext = viewModel;
-    }
-
-    protected override async void OnBindingContextChanged()
-    {
-        base.OnBindingContextChanged();
-
-        await viewModel.InitializeAsync();
-    }
+        InitializeComponent();       
+    }    
 }
